clean structure
As discussed:
models: basic model classes
options: Optional classes to be added to a network class
network: solph network classes
blocks: building constraints, set, variables... for network classes
plumbing: tools
__init__: all classes should be import as solph.classname
