+++ oemof/solph/__init__.py
@@ -4,5 +4,7 @@
 """
 from oemof.solph.network import (Sink, Source, LinearTransformer, Storage, Bus,
                                  Flow, EnergySystem)
-from oemof.solph.plumbing import OperationalModel
+from oemof.solph.models import OperationalModel
 from oemof.solph.groupings import GROUPINGS
+from oemof.solph.options import (Investment, Discrete)
+from oemof.solph.plumbing import NodesFromCSV
